refactor: use source sets instead of runtime checks
